Move loginStrategy into authService
