+++ src/commands/login.command.spec.ts
@@ -1,6 +1,5 @@
 import { mock, MockProxy } from "jest-mock-extended";
 
-import { AuthResult } from "@/jslib/common/src/models/domain/authResult";
 import { ApiLogInCredentials } from "@/jslib/common/src/models/domain/logInCredentials";
 
 import { AuthService } from "../abstractions/auth.service";
@@ -38,8 +37,6 @@ describe("LoginCommand", () => {
     process.env.BW_CLIENTID = clientId;
     process.env.BW_CLIENTSECRET = clientSecret;
 
-    authService.logIn.mockResolvedValue(new AuthResult()); // logging in with api key does not set any flag on the authResult
-
     const result = await loginCommand.run();
 
     expect(authService.logIn).toHaveBeenCalledWith(new ApiLogInCredentials(clientId, clientSecret));
@@ -52,8 +49,6 @@ describe("LoginCommand", () => {
   });
 
   it("uses client id and secret prompted from the user", async () => {
-    authService.logIn.mockResolvedValue(new AuthResult()); // logging in with api key does not set any flag on the authResult
-
     const result = await loginCommand.run();
 
     expect(authService.logIn).toHaveBeenCalledWith(new ApiLogInCredentials(clientId, clientSecret));
